+++ NEWS.md
@@ -12,6 +12,7 @@
   - -T --test option becomes -d --dry-run
 - Performance
   - Archives skipped without unpacking contents if all contents are skippable.
+  - Zip archives rewritten in place on disk if possible.
   - animated pngs better optimized with oxipng
   - Faster scanning for legacy timestamps on startup.
 - Dev